Extract logic from scraper.rb into Members class
Member names will require some additional parsing. Making the
scraper class based in preparation of that.
