+++ scraper.rb
@@ -2,11 +2,9 @@
 # encoding: utf-8
 
 require 'scraperwiki'
-require 'nokogiri'
 require 'colorize'
+require_relative 'lib/members'
 require 'pry'
-require 'open-uri/cached'
-OpenURI::Cache.cache_path = '.cache'
 
 class String
   def tidy
@@ -14,32 +12,8 @@ class String
   end
 end
 
-def noko_for(url)
-  Nokogiri::HTML(open(url).read)
-end
+URL = 'http://www.palemene.ws/new/members-of-parliament/members-of-the-xvi-parliament/'.freeze
 
-def scrape_list(url)
-  warn url
-  noko = noko_for(url)
-  noko.xpath('//div[@class="entry"]/table[1]/tbody/tr').each do |a|
-    name_and_party = a.xpath('td/text()')[1].text
-    party_id = name_and_party.split('(')[-1].gsub(')','')
-    data = {
-      name: name_and_party.split('(')[0].tidy,
-      party_id: party_id,
-      party_name: party_list[party_id],
-      area: a.xpath('td/text()')[2].text.tidy,
-      term: 16
-    }
-    ScraperWiki.save_sqlite([:name], data)
-  end
+Members.new(URL).to_a.each do |member|
+  ScraperWiki.save_sqlite([:name], member)
 end
-
-def party_list
-  {
-    'H' => 'HRPP',
-    'TS' => 'Tautua Samoa'
-  }
-end
-
-scrape_list('http://www.palemene.ws/new/members-of-parliament/members-of-the-xvi-parliament/')
